fix: utils: fix installing deb issue when using soft links.
During installation, the process looks for the parent directory of the real .src path. Since this directory defaults to requiring root permissions, it also needs to be mounted into the container. Additionally, when generating bsp_makedeb, the script copies deb packages to a specific path that originally referenced $SCRIPT_DIR/.src/. Modified this to use $DEB_DIR, which is passed into the container via environment variables during its creation.